pyproject.toml
[project]
name = "pydantic-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.12.2",
    "jupyterlab>=4.4.7",
    "mcp>=1.13.1",
    "pydantic-ai>=1.0.1",
    "python-dotenv>=1.1.1",
]
